squoosh: refactor into multiple files, add an incredibly breif design doc.
